+++ apps/web/src/modules/create-dao/components/Allocation/Toggle.tsx
@@ -1,10 +1,7 @@
 import React from 'react'
 import { Flex } from '@zoralabs/zord'
 
-import {
-  allocationToggle,
-  allocationToggleButtonVariants,
-} from '../../../../components/Fields/styles.css'
+import { allocationToggle, allocationToggleButtonVariants } from './Toggle.css'
 
 export const Toggle = ({ on, onToggle }: { on: boolean; onToggle: () => void }) => (
   <Flex className={allocationToggle[on ? 'on' : 'off']} onClick={onToggle}>
